Update tests to use context
