+++ src/__tests__/Outline.jsx
@@ -36,10 +36,12 @@ describe('Outline', () => {
       const { func: onLoadSuccess, promise: onLoadSuccessPromise } = makeAsyncCallback();
 
       shallow(
-        <Outline
-          onLoadSuccess={onLoadSuccess}
-          pdf={pdf}
-        />
+        <Outline onLoadSuccess={onLoadSuccess} />,
+        {
+          context: {
+            pdf,
+          },
+        }
       );
 
       expect.assertions(1);
@@ -52,10 +54,12 @@ describe('Outline', () => {
       muteConsole();
 
       shallow(
-        <Outline
-          onLoadError={onLoadError}
-          pdf={failingPdf}
-        />
+        <Outline onLoadError={onLoadError} />,
+        {
+          context: {
+            pdf: failingPdf,
+          },
+        }
       );
 
       expect.assertions(1);
@@ -68,10 +72,12 @@ describe('Outline', () => {
       const { func: onLoadSuccess, promise: onLoadSuccessPromise } = makeAsyncCallback();
 
       const mountedComponent = shallow(
-        <Outline
-          onLoadSuccess={onLoadSuccess}
-          pdf={pdf}
-        />
+        <Outline onLoadSuccess={onLoadSuccess} />,
+        {
+          context: {
+            pdf,
+          },
+        }
       );
 
       expect.assertions(2);
@@ -79,10 +85,8 @@ describe('Outline', () => {
 
       const { func: onLoadSuccess2, promise: onLoadSuccessPromise2 } = makeAsyncCallback();
 
-      mountedComponent.setProps({
-        onLoadSuccess: onLoadSuccess2,
-        pdf: pdf2,
-      });
+      mountedComponent.setProps({ onLoadSuccess: onLoadSuccess2 });
+      mountedComponent.setContext({ pdf: pdf2 });
 
       // It would have been .toMatchObject if not for the fact _pdf2.pdf has no outline
       await expect(onLoadSuccessPromise2).resolves.toBe(desiredLoadedOutline2);
