+++ src/__tests__/Page.jsx
@@ -52,8 +52,12 @@ describe('Page', () => {
         <Page
           onLoadSuccess={onLoadSuccess}
           pageIndex={0}
-          pdf={pdf}
-        />
+        />,
+        {
+          context: {
+            pdf,
+          }
+        }
       );
 
       expect.assertions(1);
@@ -68,8 +72,12 @@ describe('Page', () => {
       shallow(
         <Page
           onLoadError={onLoadError}
-          pdf={failingPdf}
-        />
+        />,
+        {
+          context: {
+            pdf: failingPdf,
+          }
+        }
       );
 
       expect.assertions(1);
@@ -85,8 +93,12 @@ describe('Page', () => {
         <Page
           onLoadSuccess={onLoadSuccess}
           pageIndex={0}
-          pdf={pdf}
-        />
+        />,
+        {
+          context: {
+            pdf,
+          }
+        }
       );
 
       expect.assertions(1);
@@ -102,8 +114,12 @@ describe('Page', () => {
         <Page
           onLoadSuccess={onLoadSuccess}
           pageNumber={1}
-          pdf={pdf}
-        />
+        />,
+        {
+          context: {
+            pdf,
+          }
+        }
       );
 
       expect.assertions(1);
@@ -118,9 +134,13 @@ describe('Page', () => {
       shallow(
         <Page
           pageIndex={0}
-          pdf={pdf}
           registerPage={registerPage}
-        />
+        />,
+        {
+          context: {
+            pdf,
+          }
+        }
       );
 
       expect.assertions(1);
@@ -133,9 +153,13 @@ describe('Page', () => {
       const component = shallow(
         <Page
           pageIndex={0}
-          pdf={pdf}
           unregisterPage={unregisterPage}
-        />
+        />,
+        {
+          context: {
+            pdf,
+          }
+        }
       );
 
       component.unmount();
@@ -151,8 +175,12 @@ describe('Page', () => {
         <Page
           onLoadSuccess={onLoadSuccess}
           pageIndex={0}
-          pdf={pdf}
-        />
+        />,
+        {
+          context: {
+            pdf,
+          }
+        }
       );
 
       expect.assertions(2);
@@ -181,8 +209,12 @@ describe('Page', () => {
         <Page
           className={className}
           pageIndex={0}
-          pdf={pdf}
-        />
+        />,
+        {
+          context: {
+            pdf,
+          }
+        }
       );
 
       const wrapperClassName = component.find('.react-pdf__Page').prop('className');
@@ -197,8 +229,12 @@ describe('Page', () => {
         <Page
           inputRef={inputRef}
           pageIndex={1}
-          pdf={silentlyFailingPdf}
-        />
+        />,
+        {
+          context: {
+            pdf: silentlyFailingPdf,
+          }
+        }
       );
 
       expect(inputRef).toHaveBeenCalled();
@@ -212,9 +248,13 @@ describe('Page', () => {
         <Page
           pageIndex={1}
           pageNumber={1}
-          pdf={pdf}
           onLoadSuccess={onLoadSuccess}
-        />
+        />,
+        {
+          context: {
+            pdf,
+          }
+        }
       );
 
       expect.assertions(1);
@@ -230,8 +270,12 @@ describe('Page', () => {
         <Page
           onLoadSuccess={onLoadSuccess}
           pageIndex={0}
-          pdf={pdf}
-        />
+        />,
+        {
+          context: {
+            pdf,
+          }
+        }
       );
 
       expect.assertions(1);
@@ -247,14 +291,63 @@ describe('Page', () => {
         <Page
           onLoadSuccess={onLoadSuccess}
           pageIndex={0}
-          pdf={pdf}
           rotate={90}
-        />
+        />,
+        {
+          context: {
+            pdf,
+          }
+        }
       );
 
       expect.assertions(1);
       return onLoadSuccessPromise.then(() => {
         expect(component.instance().rotate).toBe(90);
+      });
+    });
+
+    it('requests page to be rendered with default rotation when given rotate context', () => {
+      const { func: onLoadSuccess, promise: onLoadSuccessPromise } = makeAsyncCallback();
+
+      const component = shallow(
+        <Page
+          onLoadSuccess={onLoadSuccess}
+          pageIndex={0}
+        />,
+        {
+          context: {
+            pdf,
+            rotate: 90,
+          }
+        }
+      );
+
+      expect.assertions(1);
+      return onLoadSuccessPromise.then(() => {
+        expect(component.instance().rotate).toBe(90);
+      });
+    });
+
+    it('requests page to be rendered with rotation passed in props when given rotate both in props and context', () => {
+      const { func: onLoadSuccess, promise: onLoadSuccessPromise } = makeAsyncCallback();
+
+      const component = shallow(
+        <Page
+          onLoadSuccess={onLoadSuccess}
+          pageIndex={0}
+          rotate={180}
+        />,
+        {
+          context: {
+            pdf,
+            rotate: 90,
+          }
+        }
+      );
+
+      expect.assertions(1);
+      return onLoadSuccessPromise.then(() => {
+        expect(component.instance().rotate).toBe(180);
       });
     });
 
@@ -265,8 +358,12 @@ describe('Page', () => {
         <Page
           onLoadSuccess={onLoadSuccess}
           pageIndex={0}
-          pdf={pdf}
-        />
+        />,
+        {
+          context: {
+            pdf,
+          }
+        }
       );
 
       expect.assertions(1);
@@ -284,9 +381,13 @@ describe('Page', () => {
         <Page
           onLoadSuccess={onLoadSuccess}
           pageIndex={0}
-          pdf={pdf}
           renderMode="canvas"
-        />
+        />,
+        {
+          context: {
+            pdf,
+          }
+        }
       );
 
       expect.assertions(1);
@@ -304,9 +405,13 @@ describe('Page', () => {
         <Page
           onLoadSuccess={onLoadSuccess}
           pageIndex={0}
-          pdf={pdf}
           renderMode="svg"
-        />
+        />,
+        {
+          context: {
+            pdf,
+          }
+        }
       );
 
       expect.assertions(1);
@@ -324,8 +429,12 @@ describe('Page', () => {
         <Page
           onLoadSuccess={onLoadSuccess}
           pageIndex={0}
-          pdf={pdf}
-        />
+        />,
+        {
+          context: {
+            pdf,
+          }
+        }
       );
 
       expect.assertions(1);
@@ -343,9 +452,13 @@ describe('Page', () => {
         <Page
           onLoadSuccess={onLoadSuccess}
           pageIndex={0}
-          pdf={pdf}
           renderTextLayer
-        />
+        />,
+        {
+          context: {
+            pdf,
+          }
+        }
       );
 
       expect.assertions(1);
@@ -363,9 +476,13 @@ describe('Page', () => {
         <Page
           onLoadSuccess={onLoadSuccess}
           pageIndex={0}
-          pdf={pdf}
           renderTextLayer={false}
-        />
+        />,
+        {
+          context: {
+            pdf,
+          }
+        }
       );
 
       expect.assertions(1);
@@ -383,10 +500,14 @@ describe('Page', () => {
         <Page
           onLoadSuccess={onLoadSuccess}
           pageIndex={0}
-          pdf={pdf}
           renderMode="svg"
           renderTextLayer
-        />
+        />,
+        {
+          context: {
+            pdf,
+          }
+        }
       );
 
       expect.assertions(1);
@@ -404,8 +525,12 @@ describe('Page', () => {
         <Page
           onLoadSuccess={onLoadSuccess}
           pageIndex={0}
-          pdf={pdf}
-        />
+        />,
+        {
+          context: {
+            pdf,
+          }
+        }
       );
 
       expect.assertions(1);
@@ -423,9 +548,13 @@ describe('Page', () => {
         <Page
           onLoadSuccess={onLoadSuccess}
           pageIndex={0}
-          pdf={pdf}
           renderAnnotations
-        />
+        />,
+        {
+          context: {
+            pdf,
+          }
+        }
       );
 
       expect.assertions(1);
@@ -443,9 +572,13 @@ describe('Page', () => {
         <Page
           onLoadSuccess={onLoadSuccess}
           pageIndex={0}
-          pdf={pdf}
           renderAnnotations={false}
-        />
+        />,
+        {
+          context: {
+            pdf,
+          }
+        }
       );
 
       expect.assertions(1);
